add initial shadow-cljs config
The reason the output directory is stored inside of the .parcel-cache directory is because we want to disable file watcher notifications for any of the compiled JS files from shadow-cljs. This allows for shadow-cljs to manage HMR for all the clojurescript files while allowing parcel to bundle those together with other vanilla JS or CSS files.